Augment getQuotedString for MySQL
Because MySQL use 'quoted_string'@'quoted_string' in some places,
and `quoted_string`.`quoted_string` in others, while it also really don't
likes spaces to be inserted, the easiest solution appears to be to more
aggressively collect quoted strings as groups of quoted strings,
separated by @ or . characters.

This patch applies that strategy.

The `string`@bare_string might need to be added and handled, but I'm
waiting to find an actual instance where it is reasonable before I fix
it.